New test: implicitode.
This test will challenge floating point performance in the presence
of polymorphism, function parameters and other "realistic" library
situations. Hopefully the code is complex enough to defeat trivial
"unboxing" heuristics in languages where floating points are second
class citizen. This test also requires long doubles. If they are not
available, we will require that "comparable precision" be achieved.

As of now, the c++ and python files aren't the same test. The c++ test
is an emasculated version of the python one. I'm trying to keep the test
at 100 lines of c++. The python test is the test I really wanted to have.
They will be reconciled later.

Sebastien Loisel